Split esmvalcore/_recipe.py
